[new] Added uptime command
